refactor: update user component imports to use shadcn directory
