+++ features/user/components/profile-cards.tsx
@@ -5,7 +5,7 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
+} from '@/components/shadcn/card';
 import { cn } from '@/lib/utils';
 
 interface ProfileCardProps {
