+++ features/user/components/profile-edit-form.tsx
@@ -6,29 +6,29 @@ import { useUpdateUserWithFiles } from '@/hooks/user/use-user-mutations';
 import { useDeleteAttachment } from '@/hooks/attachment/use-attachment-mutations';
 import { useQueryClient } from '@tanstack/react-query';
 import { userKeys } from '@/hooks/user/user-keys';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/shadcn/button';
 import {
   RemoveProfilePictureDialog,
   RemoveCvDialog,
   SaveProfileDialog,
 } from './user-alert-dialogs';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Textarea } from '@/components/ui/textarea';
+import { Input } from '@/components/shadcn/input';
+import { Label } from '@/components/shadcn/label';
+import { Textarea } from '@/components/shadcn/textarea';
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
+} from '@/components/shadcn/select';
 import {
   Card,
   CardContent,
   CardDescription,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
+} from '@/components/shadcn/card';
 import {
   Loader2,
   Save,
